packages/pyproject.toml
[project]
name = "packages"
version = "0.1.0"
description = "Shared database ORM and models for MCP Gateway Registry"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=9.0.1",
    "weaviate-client>=4.18.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"packages" = "."}
packages = ["packages", "packages.db", "packages.db.core", "packages.db.managers", "packages.db.models", "packages.db.search"]
